## Roles **Fixes** - Witch and occultist now show proper status ("Hexed") **Changes** - Mystic now has a modifier icon
## Templates
**Trigger Fixes**
- ${TriggerMeta} for TaskComplete now shows the proper amount of completed tasks
## General
**Fixes**
- Players no longer have an initial GA shield / GA shield on the start of each round
- Players can no longer visually receive multiple of the same status,

## Roles
**Fixes**
- Multiple roles were broken due to a small bug in my implementation
- Certain indicators were not showing in meeting
- The oblivious modifier cannot be given multiple times

## Uncertain
**Fixes**
- Possibly fixed being kicked when voting out an NK (I tested this with my phone and no kicks)

+++ src/Patches/AssignRoleOnDeathPatch.cs
@@ -18,7 +18,7 @@ public class AssignRoleOnDeathPatch
     public static void Postfix(RoleManager __instance, [HarmonyArgument(0)] PlayerControl player, [HarmonyArgument(1)] bool specialRolesAllowed)
     {
         RpcV3.Immediate(player.NetId, RpcCalls.SetRole)
-            .Write((byte)(player.GetVanillaRole().IsImpostor() ? RoleTypes.ImpostorGhost : RoleTypes.CrewmateGhost))
+            .Write((ushort)(player.GetVanillaRole().IsImpostor() ? RoleTypes.ImpostorGhost : RoleTypes.CrewmateGhost))
             .Send(player.GetClientId());
         player.Data.DefaultOutfit.PetId = "pet_EmptyPet";
         VentLogger.Debug($"Dead Player {player.name} => {player.Data.Role.Role}");
